feat(notification): refactor Kafka integration and update email publishing logic
