+++ apps/notification/src/app.module.ts
@@ -3,7 +3,7 @@ import { Module } from '@nestjs/common';
 import { ScheduleModule } from '@nestjs/schedule';
 import { HttpClientModule } from 'libs/infrastructure/http/http-client.module';
 
-import { KafkaPublisherModule } from './kafka/kafka.module';
+import { KafkaPublisherModule } from './kafka/kafka.publisher.module';
 import { NotificationModule } from './notification/notification.module';
 import { SubscriptionModule } from './subscription/subscription.module';
 import { WeatherModule } from './weather/weather.module';
