--- apps/notification/src/kafka/kafka.module.ts
@@ -1,32 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ClientsModule, Transport } from '@nestjs/microservices';
-import { KafkaConfig } from 'apps/notification/config/kafka.config';
-
-export const KAFKA_PUBLISHER = Symbol('KAFKA_PUBLISHER');
-
-@Module({
-  imports: [
-    ClientsModule.registerAsync([
-      {
-        name: KAFKA_PUBLISHER,
-        useFactory: (config: KafkaConfig) => ({
-          transport: Transport.KAFKA,
-          options: {
-            client: {
-              clientId: config.clientId,
-              brokers: [`${config.host}:${config.port}`],
-            },
-            producer: {
-              allowAutoTopicCreation: true,
-            },
-            producerOnlyMode: true,
-          },
-        }),
-
-        inject: [KafkaConfig],
-      },
-    ]),
-  ],
-  exports: [ClientsModule],
-})
-export class KafkaPublisherModule {}
