+++ apps/notification/src/kafka/kafka.publisher.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+
+import { kafkaProvider } from './kafka.producer';
+
+@Module({
+  providers: [kafkaProvider],
+  exports: [kafkaProvider],
+})
+export class KafkaPublisherModule {}
